Add onnx and ort to test requirements
