+++ test_requirements.txt
@@ -5,3 +5,5 @@ h5py
 pillow
 pkbar
 pyyaml
+onnx 
+onnxruntime
