+++ torchani/onnx/__init__.py
@@ -1,1 +1,2 @@
-from .onnx_workarounds import *  # noqa
+from .aev_workarounds import *  # noqa
+from .animodel_workarounds import *  # noqa
